workplace-search/introduction_workshop/requirements.txt
elastic-enterprise-search==7.11.0
PyYAML>=5.3.1
